​refactor: 유틸 인풋 분리(view 폴더 ➔ utils 폴더)
