+++ src/utils/input.js
@@ -1,0 +1,27 @@
+import readline from "readline";
+
+export const readLineAsync = (query = "") => {
+  return new Promise((resolve) => {
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(query, (input) => {
+      rl.close();
+      resolve(input);
+    });
+  });
+};
+
+export const retryUntilValid = async (getInputFn, validator) => {
+  while (true) {
+    try {
+      const userInput = await getInputFn();
+      validator(userInput);
+      return userInput;
+    } catch (error) {
+      console.error(error.message);
+    }
+  }
+};
